external working

+++ Laravel/app/Http/Controllers/PartnerTrainingUserController.php
@@ -3,10 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Material;
+use App\MaterialPartnerTrainingUser;
 use App\Partner;
 use App\PartnerTrainingUser;
-use App\Role;
-
 use App\Training;
 use App\User;
 use Illuminate\Http\Request;
@@ -14,11 +13,7 @@ use Illuminate\Support\Facades\DB;
 
 class PartnerTrainingUserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
         $partner_Training_Users = PartnerTrainingUser::with('partner', 'training', 'user')->get();
@@ -41,15 +36,10 @@ class PartnerTrainingUserController extends Controller
         return view('external.show', compact('partner_Training_Users'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
         $partner_Training_Users = PartnerTrainingUser::all();
-        //$role_users = Role_User::all();
         $partners=Partner::all();
         $users=User::all();
         $trainings=Training::all();
@@ -111,22 +101,14 @@ class PartnerTrainingUserController extends Controller
 
 
 
-        //$materials = Material::where('isInternal', '=', false)->get();
 
         $materials = Material::with('partnerTrainingUsers')->where('isInternal', false)->get();
-        //dd($materials)      ;
 
         return view('external.edit', compact('partner_Training_Users', 'partners', 'trainings', 'users',   'materials' ));
     }
 
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\PartnerTrainingUsers  $partner_Training_Users
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
         $partner_Training_Users = PartnerTrainingUser::with('partner', 'training', 'user')->findOrFail($id);
@@ -142,37 +124,36 @@ class PartnerTrainingUserController extends Controller
         if (strtotime($request->start_date) > strtotime($request->end_date)) {
             return redirect()->back()->withErrors(['end_date' => 'End date must be after or equal to start date']);
         }
+
         $partner_Training_Users->update($request->all());
 
         $selectedMaterials = $request->input('materials');
         $materialQuantities = $request->input('material_quantities');
+
         if ($selectedMaterials) {
             foreach ($selectedMaterials as $materialId) {
                 $quantity = $materialQuantities[$materialId] ?? 1;
 
-                $partner_Training_Users->materials()->syncWithoutDetaching([
-                    $materialId => ['quantity' => $quantity],
-                ]);
+                $currentQuantity = $partner_Training_Users->materials->where('id', $materialId)->first()->pivot->quantity ?? 0;
+                $quantityDecreased = $currentQuantity > $quantity;
+
+                if ($quantity > 0) {
+                    $partner_Training_Users->materials()->syncWithoutDetaching([
+                        $materialId => ['quantity' => $quantity],
+                    ]);
+                } else {
+                    // If the quantity is 0, delete the pivot record
+                    $partner_Training_Users->materials()->detach($materialId);
+                }
 
                 $material = Material::find($materialId);
 
                 if ($material) {
-                    $material->decrement('quantity', $quantity);
+                    $material->increment('quantity', $quantityDecreased ? ($currentQuantity - $quantity) : 0);
+                    $material->decrement('quantity', $quantityDecreased ? 0 : $quantity);
                 }
             }
         }
-
-//        $partner_Training_Users->materials()->detach();
-//
-//        foreach ($request['material_quantities'] as $materialId => $quantity) {
-//            $partner_Training_Users->materials()->attach($materialId, ['quantity' => $quantity]);
-//
-//        }
-       // dd($partner_Training_Users->materials()->attach($materialId, ['quantity' => $quantity]));
-
-
-
-
 
         return redirect()->route('external.index')->with('success', 'Formação atualizada com sucesso');
     }
@@ -182,21 +163,24 @@ class PartnerTrainingUserController extends Controller
 
 
 
+
+
     public function destroy($id)
     {
-        $partnerTrainingUser = PartnerTrainingUser::findOrFail($id);
 
+        $partnerTrainingUser = PartnerTrainingUser::find($id);
 
-        $materialTrainings = $partnerTrainingUser->Material_Training;
-
-
-        foreach ($materialTrainings as $materialTraining) {
-            $material = $materialTraining->material;
-            $material->quantity += $materialTraining->quantity;
-            $material->save();
+        if (!$partnerTrainingUser) {
+            return redirect()->route('partner-training-users.index')->with('error', 'PartnerTrainingUser not found.');
         }
 
-        $partnerTrainingUser->Material_Training()->delete();
+        $materials = $partnerTrainingUser->materials;
+
+        MaterialPartnerTrainingUser::where('partner_training_user_id', $partnerTrainingUser->id)->delete();
+
+        foreach ($materials as $material) {
+            $material->increment('quantity', $material->pivot->quantity);
+        }
 
         $partnerTrainingUser->delete();
 
@@ -205,17 +189,30 @@ class PartnerTrainingUserController extends Controller
 
 
 
+
     public function massDelete(Request $request)
     {
-
-
-
         $request->validate([
             'ptu_ids' => 'required|array',
-            'ptu_ids.*' => 'exists:partner__training__users,id',]);
+            'ptu_ids.*' => 'exists:partner_training_users,id',
+        ]);
 
-        //dd($request->all());
         try {
+            $partnerTrainingUsers = PartnerTrainingUser::whereIn('id', $request->input('ptu_ids'))->get();
+
+            foreach ($partnerTrainingUsers as $partnerTrainingUser) {
+                $materials = $partnerTrainingUser->materials;
+
+                foreach ($materials as $material) {
+                    $quantity = $material->pivot->quantity ?? 0;
+
+                    $material->increment('quantity', $quantity);
+                }
+
+                MaterialPartnerTrainingUser::where('partner_training_user_id', $partnerTrainingUser->id)->delete();
+            }
+
+            // Delete PartnerTrainingUsers
             PartnerTrainingUser::whereIn('id', $request->input('ptu_ids'))->delete();
 
             return redirect()->back()->with('success', 'Formações selecionadas excluídas com sucesso!');
@@ -223,6 +220,8 @@ class PartnerTrainingUserController extends Controller
             return redirect()->back()->with('error', 'Erro ao excluir Formações selecionadas. Por favor, tente novamente.');
         }
     }
+
+
